e2e/requirements.txt
pytest==8.3.5
pytest-playwright==0.7.0
pytest-html==4.1.1
requests==2.32.3
